Move Elixir settings into the top-level Elixir configurable
The settings were previously on a separate "Experimental Settings" child
page. Move them directly into the main Elixir settings page so users
don't need to navigate to a child page.

- Move panel, apply/reset/isModified logic into ElixirTopLevelConfigurable
- Remove ElixirExperimentalSettingsConfigurable and its plugin.xml
  registration
- Update searchable option contributor to point at the main configurable
